Initial import of working code
Designed to be simple and correct. No thought put in to optimization, yet.
Correctly identifies valid UTF-8 sequences.
Shows strings of 4 *characters* or longer. (Not bytes).
Works on stdin or a single filename may be specified.

It works for my purposes and probably is fine for others as well.

Deficiencies: Hardcoded to strings of minlength 4.
	      Could be a lot faster with some simple optimizations.
	      Does not handle any options.
	      Should be merged with `strings` from GNU binutils.